restructure model structure
