+++ face_alignment/api.py
@@ -1,15 +1,9 @@
-import os
 import torch
 from torch.utils.model_zoo import load_url
 from enum import Enum
 from skimage import io
 from skimage import color
 import numpy as np
-import cv2
-try:
-    import urllib.request as request_file
-except BaseException:
-    import urllib as request_file
 
 from .models import FAN, ResNetDepth
 from .utils import *
