+++ face_alignment/models/__init__.py
@@ -1,0 +1,2 @@
+from .fan import FAN
+from .resnet import ResNetDepth
